[webkitcorepy] run-webkit-tests may hang with python2 after r271683 https://bugs.webkit.org/show_bug.cgi?id=227715
Reviewed by Philippe Normand.

When an exception is raised from a worker if the workers are terminated
via a SIGKILL signal that later causes the task-queue to hang at close().
The issue is not reproducible with python3. So keep the same behaviour for
python3, and for python2 just use terminate() to stop the workers.

* Scripts/libraries/webkitcorepy/webkitcorepy/task_pool.py:
(TaskPool.__exit__):


Canonical link: https://commits.webkit.org/239544@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279775 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/libraries/webkitcorepy/webkitcorepy/task_pool.py
@@ -458,7 +458,8 @@ class TaskPool(object):
 
                 if sys.version_info >= (3, 7):
                     worker.kill()
-                elif hasattr(signal, 'SIGKILL'):
+                # With python2 killing directly the workers causes the queue to hang on close() <https://webkit.org/b/227715>
+                elif hasattr(signal, 'SIGKILL') and sys.version_info.major > 2:
                     try:
                         os.kill(worker.pid, signal.SIGKILL)
                     except OSError as e:
